add WholeStageResultIterator
