+++ cpp/compile.sh
@@ -2,17 +2,17 @@
 
 set -exu
 
-BUILD_TYPE=release
+BUILD_TYPE=debug
 BUILD_GAZELLE_CPP_BACKEND=OFF
-BUILD_VELOX_BACKEND=OFF
+BUILD_VELOX_BACKEND=ON
 BUILD_TESTS=OFF
-BUILD_BENCHMARKS=OFF
-BUILD_JEMALLOC=ON
+BUILD_BENCHMARKS=ON
+BUILD_JEMALLOC=OFF
 ENABLE_HBM=OFF
 BUILD_PROTOBUF=OFF
 ENABLE_S3=OFF
 ENABLE_HDFS=OFF
-NPROC=$(nproc --ignore=2)
+NPROC=4
 ARROW_ROOT=
 VELOX_HOME=
 
@@ -118,4 +118,4 @@ cmake .. \
   -DBUILD_PROTOBUF=${BUILD_PROTOBUF} \
   -DVELOX_ENABLE_S3=${ENABLE_S3} \
   -DVELOX_ENABLE_HDFS=${ENABLE_HDFS}
-make -j$NPROC
+make -j$NPROC generic_benchmark
