+++ cpp/velox/benchmarks/GenericBenchmark.cc
@@ -20,7 +20,7 @@
 #include <benchmark/benchmark.h>
 #include <gflags/gflags.h>
 #include <operators/writer/ArrowWriter.h>
-#include <velox/exec/PlanNodeStats.h>
+//#include <velox/exec/PlanNodeStats.h>
 #include <velox/exec/Task.h>
 
 #include <chrono>
@@ -100,7 +100,7 @@ auto BM_Generic = [](::benchmark::State& state,
           return sum + iter->GetCollectBatchTime();
         });
 
-    auto* rawIter = static_cast<gluten::WholeStageResIter*>(resultIter->GetRaw());
+    auto* rawIter = static_cast<gluten::WholeStageResultIterator*>(resultIter->GetRaw());
     const auto& task = rawIter->task_;
     const auto& planNode = rawIter->planNode_;
     auto statsStr = ::facebook::velox::exec::printPlanWithStats(*planNode, task->taskStats(), true);
